// MovieApp/src/types/homeType.ts
export interface HomeType {
    page: number;
    results: Result[];
    total_pages: number;
    total_results: number;
}

export interface Result {
    adult: boolean;
    backdrop_path: string;
    first_air_date: Date;
    id: number;
    media_type: string;
    name: string;
    origin_country: any[];
    original_language: string;
    original_name: string;
    overview: string;
    popularity: number;
    poster_path: string;
    vote_average: number;
    vote_count: number;
}